+++ eoreader/products/optical/sv1_product.py
@@ -42,6 +42,7 @@ from eoreader.bands import BandNames, SpectralBand
 from eoreader.bands import spectral_bands as spb
 from eoreader.exceptions import InvalidProductError
 from eoreader.products import VhrProduct
+from eoreader.products.optical.optical_product import RawUnits
 from eoreader.stac import GSD, ID, NAME, WV_MAX, WV_MIN
 from eoreader.utils import DATETIME_FMT, EOREADER_NAME, simplify
 
@@ -122,6 +123,7 @@ class Sv1Product(VhrProduct):
         self._ms_res = 2.0
         self.needs_extraction = False
         self._proj_prod_type = [Sv1ProductType.L1B]
+        self._raw_units = RawUnits.DN
 
         # Post init done by the super class
         super()._pre_init(**kwargs)